requirements.txt
imageio==2.8.0
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.2.1
numpy==1.18.3
opencv-contrib-python==4.1.2.30
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
sklearn==0.0
tensorboard==2.2.1
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0